Move /start_vote to vote.py and remove delete_log.py
